[general] Improved version info generation.
We store the coq version information in a generated ml file. This way,
dependencies are properly tracked by the users of such information,
and properly kept up to date on minor deltas.

We place the generate file on its own directory/module as to allow
make to work at finer level (assuming a future where directory ==
module).

For now the date stored is the last date of the index change, this
could be refined even more if desired.