Remove the deprecation for some 8.2-8.5 compatibility notations.
As was decided during the WG.
The notations we keep deprecating are the ones where the difference
is only a prefix becoming a qualified module name.
The intention is to turn the warning for deprecated notations on.
We change the compat version to 8.6 to allow the removal of VOld and V8_5.